Trabalhos Práticos/Trabalho Prático 1/sistema-academia/requirements.txt
streamlit
mysql-connector-python
pandas
python-dotenv
plotly